// migrations/snapshots/acd87359a4504a8b36cddc45bd0ac416c6089c43b2d424d579b017577c3ff5ec/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Transaction: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly amount: CodecTypes['pg/numeric@1']['output'];
      readonly type: 'income' | 'expense';
      readonly category:
        | 'Salary'
        | 'Food'
        | 'Transport'
        | 'Shopping'
        | 'Bills'
        | 'Entertainment'
        | 'Health'
        | 'Other';
      readonly date: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Transaction: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly amount: CodecTypes['pg/numeric@1']['input'];
      readonly type: 'income' | 'expense';
      readonly category:
        | 'Salary'
        | 'Food'
        | 'Transport'
        | 'Shopping'
        | 'Bills'
        | 'Entertainment'
        | 'Health'
        | 'Other';
      readonly date: CodecTypes['pg/text@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly transaction: {
      readonly amount: CodecTypes['pg/numeric@1']['output'];
      readonly category:
        | 'Salary'
        | 'Food'
        | 'Transport'
        | 'Shopping'
        | 'Bills'
        | 'Entertainment'
        | 'Health'
        | 'Other';
      readonly date: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly type: 'income' | 'expense';
      readonly userId: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly transaction: {
      readonly amount: CodecTypes['pg/numeric@1']['input'];
      readonly category:
        | 'Salary'
        | 'Food'
        | 'Transport'
        | 'Shopping'
        | 'Bills'
        | 'Entertainment'
        | 'Health'
        | 'Other';
      readonly date: CodecTypes['pg/text@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly type: 'income' | 'expense';
      readonly userId: CodecTypes['pg/text@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;